Remove dates from URLs and use modified dates.
I don't want folks to get the impression that content that was written
a long time ago is outdated. To help with this, I am removing dates from
the URLs. Dates will remain in the posts.

Additionally, posts can be marked with the date they were last reviewed
to help with the "is this content stale?" impression.